Moving ClientBuilder and SpApiRoster into Utilities namespace
