+++ src/Container/SpApi.php
@@ -97,11 +97,12 @@ use Glue\SpApi\OpenAPI\Clients\TokensV20210301\Model\CreateRestrictedDataTokenRe
 use Glue\SpApi\OpenAPI\Exceptions\DomainApiException;
 use Glue\SpApi\OpenAPI\Exceptions\LwaAccessTokenException;
 use Glue\SpApi\OpenAPI\Exceptions\RestrictedDataTokenException;
-use Glue\SpApi\OpenAPI\Services\Builder\ClientBuilder;
 use Glue\SpApi\OpenAPI\Services\Factory\ClientFactoryInterface;
 use Glue\SpApi\OpenAPI\Services\Lwa\LwaServiceInterface;
 use Glue\SpApi\OpenAPI\Services\Rdt\RdtServiceInterface;
 use Glue\SpApi\OpenAPI\SpApiConfig;
+use Glue\SpApi\OpenAPI\Utilities\ClientBuilder;
+use Glue\SpApi\OpenAPI\Utilities\SpApiRoster;
 use GuzzleHttp\Psr7\Stream;
 
 class SpApi implements SpApiInterface
