+++ src/Services/Factory/ClientFactory.php
@@ -54,8 +54,8 @@ use Glue\SpApi\OpenAPI\Clients\VendorTransactionStatusV1\Api\VendorTransactionAp
 use Glue\SpApi\OpenAPI\Exceptions\LwaAccessTokenException;
 use Glue\SpApi\OpenAPI\Exceptions\RestrictedDataTokenException;
 use Glue\SpApi\OpenAPI\Services\Authenticator\ClientAuthenticatorInterface;
-use Glue\SpApi\OpenAPI\Services\Builder\ClientBuilder;
 use Glue\SpApi\OpenAPI\SpApiConfig;
+use Glue\SpApi\OpenAPI\Utilities\ClientBuilder;
 
 class ClientFactory implements ClientFactoryInterface
 {
