+++ tests/Utilities/ClientBuilderTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Services;
+namespace Tests\Utilities;
 
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Api\OrdersV0Api;
 use Glue\SpApi\OpenAPI\Clients\OrdersV0\Configuration;
 use Glue\SpApi\OpenAPI\Exceptions\ClientBuilderException;
 use Glue\SpApi\OpenAPI\Services\Authenticator\ClientAuthenticatorInterface;
-use Glue\SpApi\OpenAPI\Services\Builder\ClientBuilder;
 use Glue\SpApi\OpenAPI\SpApiConfig;
+use Glue\SpApi\OpenAPI\Utilities\ClientBuilder;
 use GuzzleHttp\Client;
 use Mockery\MockInterface;
 use Tests\TestCase;
